A QnD fix to answer which files to process.
Currently executes an arbitrary file in the current working directory.
This requires each project then to define said file. A longer term
solution should use regexes and/or globs and warn about unignored files.
I digress... For now this should provide you with the flexibility to
define the rules however you want to find files that should be
processed.